Moved Firebase grabbing to ViewModel in ListEmote
