+++ source/app/src/main/java/com/example/emote/ui/addemote/AddEmoteFragment.java
@@ -12,6 +12,7 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
+import com.example.emote.Emotion;
 import com.example.emote.FireStoreHandler;
 import com.example.emote.R;
 import com.example.emote.EmotionEvent;
@@ -39,9 +40,9 @@ public class AddEmoteFragment extends Fragment {
 
 
         FireStoreHandler fsh = new FireStoreHandler("john123");
-        EmotionEvent emoteEvent1 = new EmotionEvent("Happy", Situation.FEW_PEOPLE, "Good food", new Date(System.currentTimeMillis()));
-        EmotionEvent emoteEvent2 = new EmotionEvent("Sad", Situation.FEW_PEOPLE, "Dog died", new Date(System.currentTimeMillis()));
-        EmotionEvent emoteEvent3 = new EmotionEvent("Tired", Situation.FEW_PEOPLE, "Sleepy", new Date(System.currentTimeMillis()));
+        EmotionEvent emoteEvent1 = new EmotionEvent(Emotion.HAPPY, Situation.FEW_PEOPLE, "Good food", new Date(System.currentTimeMillis()));
+        EmotionEvent emoteEvent2 = new EmotionEvent(Emotion.ANNOYED, Situation.FEW_PEOPLE, "Dog died", new Date(System.currentTimeMillis()));
+        EmotionEvent emoteEvent3 = new EmotionEvent(Emotion.EXCITED, Situation.FEW_PEOPLE, "Sleepy", new Date(System.currentTimeMillis()));
         fsh.addEmote(emoteEvent1);
         fsh.addEmote(emoteEvent2);
         fsh.addEmote(emoteEvent3);
