+++ source/app/src/main/java/com/example/emote/ui/listemote/EmoteListAdapter.java
@@ -42,7 +42,7 @@ public class EmoteListAdapter extends ArrayAdapter<EmotionEvent> {
         TextView emoteTextView = view.findViewById(R.id.emote_text);
         TextView userTextView = view.findViewById(R.id.user_name);
         TextView dateTextView = view.findViewById(R.id.date);
-        emoteTextView.setText(emote.getEmote());
+        emoteTextView.setText(emote.getEmote().toString());
         userTextView.setText(emote.getUsername());
         try{
             DateFormat df = DateFormat.getDateInstance(DateFormat.SHORT);
