+++ source/app/src/main/java/com/example/emote/ui/listemote/ListEmoteViewModel.java
@@ -1,19 +1,56 @@
 package com.example.emote.ui.listemote;
 
+import android.util.Log;
+
+import androidx.annotation.NonNull;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
+import com.example.emote.EmotionEvent;
+import com.example.emote.FireStoreHandler;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
+import com.google.firebase.firestore.QuerySnapshot;
+
+import java.util.ArrayList;
+
+import static androidx.constraintlayout.widget.Constraints.TAG;
+
 public class ListEmoteViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
+    private FireStoreHandler fsh;
+    private FirebaseFirestore db;
 
     public ListEmoteViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is List Emote fragment");
+        fsh = new FireStoreHandler("john123");
+        db = fsh.getFireStoreDBReference();
     }
 
-    public LiveData<String> getText() {
-        return mText;
+
+    public void grabFirebase(final EmoteListAdapter adapter, final ArrayList<EmotionEvent> emoteDataList){
+        // TODO: Move this to FireStoreHandler
+        db.collection(FireStoreHandler.EMOTE_COLLECTION)
+                .whereEqualTo(EmotionEvent.USERNAME_KEY, fsh.getUsername())
+                .get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        ArrayList<EmotionEvent> new_emotes;
+                        if (task.isSuccessful()) {
+                            emoteDataList.clear();
+                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                Log.d(TAG, document.getId() + " => " + document.getData());
+                                emoteDataList.add(document.toObject(EmotionEvent.class));
+                            }
+                            adapter.notifyDataSetChanged();
+                        } else {
+                            Log.d(TAG, "Error getting documents: ", task.getException());
+                        }
+                    }
+                });
     }
 }
